Added lang to html tag
