+++ pages/_document.tsx
@@ -1,6 +1,7 @@
 import { Stylesheet, InjectionMode } from '@uifabric/merge-styles';
 import { resetIds } from '@uifabric/utilities';
 import Document, { DocumentContext, Html, Head, Main, NextScript } from 'next/document';
+import { isNavigatorLanguageItalian } from '../services/Utils';
 
 // Do this in file scope to initialize the stylesheet before Fluent UI React components are imported.
 const stylesheet = Stylesheet.getInstance();
@@ -12,21 +13,26 @@ stylesheet.setConfig({
 });
 
 // Now set up the document, and just reset the stylesheet.
-export default class MyDocument extends Document {
+export default class MyDocument extends Document<{ language: string }> {
     static async getInitialProps(ctx: DocumentContext) {
         stylesheet.reset();
         resetIds();
 
         const initialProps = await Document.getInitialProps(ctx);
+        const language = ctx.req?.headers['accept-language'];
+
         return {
             ...initialProps,
-            styles: [initialProps.styles, <style key="fluentui-css" dangerouslySetInnerHTML={{ __html: stylesheet.getRules(true) }} />],
+            language: language,
+            styles: [ initialProps.styles, <style key="fluentui-css" dangerouslySetInnerHTML={{ __html: stylesheet.getRules(true) }} /> ]
         }
     }
 
     render() {
+        const language = isNavigatorLanguageItalian(this.props.language) ? "it" : "en";
+
         return (
-            <Html prefix="og: http://ogp.me/ns#">
+            <Html prefix="og: http://ogp.me/ns#" lang={language}>
                 <Head />
                 <body>
                     <Main />
